// src/types/steps.ts
export enum EnumSteps {
  BOAS_VINDAS = 0,
  IDENTIFICACAO = 1,
  FORCAS = 2,
  FRAQUEZAS = 3,
  OPORTUNIDADES = 4,
  AMEACAS = 5,
  FINANCEIRO = 6,
  PRIORIDADES = 7,
  FINALIZACAO = 8,
  RESULTADOS = 9
}

export const STEPS = [
  { label: "Boas-vindas" },             // 0
  { label: "Identificação" },          // 1
  { label: "Forças" },                 // 2
  { label: "Fraquezas" },              // 3
  { label: "Oportunidades" },          // 4
  { label: "Ameaças" },                // 5
  { label: "Saúde Financeira" },       // 6
  { label: "Prioridades" },            // 7
  { label: "Finalização" },            // 8
  { label: "Resultados" },             // 9
];